% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_one.R
\name{add_one}
\alias{add_one}
\title{Adding one to a number}
\usage{
add_one(x)
}
\arguments{
\item{x}{A numeric value}
}
\value{
A numeric value
}
\description{
Adding one to a number
}
\examples{
add_one(5) #6
}
