% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT.R
\name{assumption}
\alias{assumption}
\title{Assumption}
\usage{
assumption(x, testtype)
}
\arguments{
\item{x}{text string; This should be the data file name of the test sample.}
}
\value{
3 graphs of linearity, constant variance and normal distribution of the residuals
}
\description{
This function outputs the 3 graphs of assumption for the linear regression model.
}
