% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{DATASET}
\alias{DATASET}
\title{Sample Data for Analysis}
\format{
A \code{tibble} with 2 columns, which are:
\describe{
 \item{x}{random normal data}
 \item{y}{outcome that is normal based on \code{x}}
 }
}
\usage{
DATASET
}
\description{
A simple \code{tibble} that is for simple testing and running of examples
}
\keyword{datasets}
