% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Visualizes the data of the given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The numeric input for state number to filter data}

\item{year}{The numeric input for year}
}
\value{
The plot of the filtered data
}
\description{
This function defines the file name with \link{make_filename}, and reads it
with \link{fars_read}. Then, extract the rows when the state label is the
given state number. It removes the data for the longitude more than 900 and
the latitude more than 90 since they would be false data. Finally it visualize
data on a plot.
}
\note{
It runs into a warning if there is no data for the defined
}
\examples{
\dontrun{
fars_map_state(1,2013)
}
}
