% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads .csv file as table}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the file name to be read}
}
\value{
This function returns to a table frame including the info from the
given filename.
}
\description{
This is a simple function to read a csv file. First, it checks if the given
filename exists. If it exists, it reads it. Then it converts the resulting
data frame to table.
}
\note{
If the file for the given year does not exist, it turns
into a warning.
}
\examples{
\dontrun{
tbl_data <- fars_read('accident_2013.csv.bz2')
}
}
