% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract month list for the selected years from data}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The numeric array for years}
}
\value{
The table frame including months for the given years from the data
}
\description{
This function creates a character string of the file name for the given year
with \link{make_filename}. Then, it reads that file with \link{fars_read}. It
adds the year as and additional field to the table frame. Finally it selects
month and year from data. This function can handle this process with multiple
inputs.
}
\note{
Also, it turns into a warning message if any error occurs. It means
the given year is not valid.
}
\examples{
fars_read_years(2013:2015)

}
