% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes the given year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The numeric array for years}
}
\value{
The summary of data of month for the given years
}
\description{
This function selects the monthly data for the selected years with
\link{fars_read_years} and groups the data by year and month. Then, it
summarize data and spread the year.
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
}
}
