% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funfun.R
\name{funfun}
\alias{funfun}
\title{Finding Powers}
\usage{
funfun(x)
}
\arguments{
\item{x}{text string; number to find powers of.}
}
\value{
Text string with powers
}
\description{
This finds the square, cube and self power of number given
}
\examples{
funfun(5)
}
