% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alfred.R
\name{get_alfred_series}
\alias{get_alfred_series}
\title{Get Alfred Series}
\usage{
get_alfred_series(series_id, series_name = NULL,
  observation_start = NULL, observation_end = NULL,
  realtime_start = NULL, realtime_end = NULL)
}
\arguments{
\item{series_id}{jj}

\item{series_name}{jj}

\item{observation_start}{jj}

\item{observation_end}{jj}

\item{realtime_start}{jj}

\item{realtime_end}{jj}
}
\value{
The output from \code{\link{print}}
}
\description{
Get Alfred Series
}
\examples{
\dontrun{
get_alfred_series("GDPC1", "rgdp", observation_start = "2007-05-31", realtime_start = "2008-05-31", realtime_end = "2009-03-30")
}

}
