% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{get_fred_series}
\alias{get_fred_series}
\title{Get Fred Series}
\usage{
get_fred_series(
  series_id,
  series_name = NULL,
  observation_start = NULL,
  observation_end = NULL
)
}
\arguments{
\item{series_id}{FRED time series ID.}

\item{series_name}{Choose the name for the series column in output dataframe. Default: series_id.}

\item{observation_start}{Dtae of the first observation in "yyyy-mm-dd" format. Default: earliest observation avaliable.}

\item{observation_end}{Date of last observation in "yyyy-mm-dd" format. Default: Last observation available.}
}
\description{
This function can pull time series from the FRED database \url{https://fred.stlouisfed.org}.
}
\examples{
\dontrun{
get_fred_series("INDPRO", "indpro")
}

}
