% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldnumber.R
\name{goldnumber}
\alias{goldnumber}
\title{Applies and prints the golden ratio to each index(n by n-1) of a Fibonacci sequence}
\usage{
goldnumber(a, b)
}
\arguments{
\item{a}{is a positive integer number, the argument used for generate a Fibonacci sequence.}

\item{b}{also a positive integer number, the argument used for determinate how many decimal
places needed for approximation.}
}
\value{
two vectors with the Fibonacci sequence and the gold ratio between n-th and the index n-1.
}
\description{
This function print a Fibonacci sequence and a sequence of ratio between the n-th and index n-1
according with the second argument that's determinate how many decimals places be needed to get the
rounded gold number.
This function show too how many ratios has been calculated for get a
approximation of the gold number with 3 decimals places.
}
\details{
This is just an initial test for constructing R packages using git and
GitHub on-line.
}
