% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodbye.R
\name{goodbye}
\alias{goodbye}
\title{Goodbye function}
\usage{
goodbye(name)
}
\arguments{
\item{name}{Insert a name, or saying}
}
\value{

}
\description{
Say goodbye how you like
}
\examples{
goodbye('Derek')
}
