% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My Hello World Function}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of the person to say}
}
\value{
The ouput from \code{\link{print}}
}
\description{
Prints Hello person
}
\examples{
hello("Gustavo")
\dontrun{
hello("Steve")
}
}
