% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_goodbye.R
\name{hello_goodbye}
\alias{hello_goodbye}
\title{Hello Goodbye}
\usage{
hello_goodbye(name)
}
\arguments{
\item{name}{Insert a name, or saying}
}
\value{
vector with salutation
}
\description{
Say hello how you like
}
\examples{
hello_goodbye('Name')
}
