% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Obtained Interarrival Time from a simulated Poisson Process}
\usage{
main(seed = 723, t = 24, lambda = 2, p = 1/3)
}
\arguments{
\item{seed}{A user determined seed to randomly generate Poisson process}

\item{t}{time of Poisson Process}

\item{lambda}{lambda of Poisson Process}

\item{p}{probability of the truncated Poisson}
}
\value{
Interarrival time from the simulated Poisson Process
}
\description{
This function provides functionality for obtaining interarrival time
from a simulated Poisson Process.
}
\examples{
main(seed=723,t=12,lambda=2,p=1)
}
