% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a string for the file name from year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The numeric input for year}
}
\value{
A character string of file name
}
\description{
This function creates the file name as string with a format of .csv.bz2. The
file name is defined as e.g. accident_2020.csv.bz2.
}
\examples{
\dontrun{
make_filename(2013)
}
}
