% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minus.R
\name{minus}
\alias{minus}
\title{Returns the differ of two numbers}
\usage{
minus(a, b)
}
\arguments{
\item{a}{Number one.}

\item{b}{Number two.}
}
\value{
Differ of two numbers
}
\description{
Returns the differ of two numbers
}
\examples{
library(tester)
minus(1, 2)
}
