% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{Silly printer function. Test #2}
\usage{
printer(r, x)
}
\arguments{
\item{r}{what you want in second column}

\item{x}{what you want in first column}
}
\value{
a tibble
}
\description{
Silly printer function. Test #2
}
\examples{
printer(x = rnorm(5), r= rnorm(5))
}
