% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_read.R
\name{sample_read}
\alias{sample_read}
\title{Read CSV duplicate}
\usage{
sample_read(path)
}
\arguments{
\item{path}{path to filename}
}
\value{
A \code{tibble}
}
\description{
Read CSV duplicate
}
\examples{
csv = system.file("extdata", "sample_data.csv", package = "tester")
sample_read(csv)
}
