% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_foo.R
\name{time_foo}
\alias{time_foo}
\title{Time function}
\usage{
time_foo(x)
}
\arguments{
\item{x}{a parameter that is returned after 1 second}
}
\value{
x
}
\description{
Time function
}
\examples{
time_foo("asd")
}
