% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_labelled.R
\name{write_labelled}
\alias{write_labelled}
\title{Writes Labelled Data into Excel File with Variable and Value Labels}
\usage{
write_labelled(x, file)
}
\arguments{
\item{x}{Labelled data frame}

\item{file}{Excel workbook with two sheets: for data and dictionary}
}
\description{
Writes Labelled Data into Excel File with Variable and Value Labels
}
\examples{
# write_labelled(pss, "example.xlsx")
}
