% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_labelled.R
\name{zap_na}
\alias{zap_na}
\title{Recodes user missing values to NA, and removes their labels}
\usage{
zap_na(x)
}
\arguments{
\item{x}{Variable which missing values will be recoded into NA}
}
\description{
Recodes user missing values to NA, and removes their labels
}
\examples{
var <- c(1,1,2,2,2,7,8,9,0)
attr(var, "na_values") <- c(7,8,9)
zap_na(var)
}
