% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_labelled.R
\name{zap_values}
\alias{zap_values}
\title{Recodes specified values to NA, also removes their labels}
\usage{
zap_values(x, values)
}
\arguments{
\item{x}{Variable which values are to be recoded into NA}

\item{values}{Vector of values to recode}
}
\description{
Recodes specified values to NA, also removes their labels
}
\examples{
var <- c(1,1,2,2,2,7,8,9,0)
zap_values(var, 7:9)
zap_values(var, c(1,2,3))
}
