% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibonacci}
\alias{fibonacci}
\title{Print Fibonacci sequence}
\usage{
fibonacci(n = 3)
}
\arguments{
\item{n}{is a natural number.}
}
\value{
A vector containing the first 'n' Fibonacci numbers.
}
\description{
This function prints the 'n' first Fibonacci numbers. Here the 'n' is
a natural number.
}
\details{
This is just an initial test for constructing R packages using git and
GitHub on-line.
}
