% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{get_fred_series}
\alias{get_fred_series}
\title{Get Fred Series}
\usage{
get_fred_series(series_id, series_name = NULL,
  observation_start = NULL, observation_end = NULL)
}
\arguments{
\item{series_id}{jj}

\item{series_name}{jj}

\item{observation_start}{jj}

\item{observation_end}{jj}
}
\value{
The output from \code{\link{print}}
}
\description{
Get Fred Series
}
\examples{
\dontrun{
get_fred_series("INDPRO", "indpro")
}

}
