% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My hello function}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of the person to say hi}
}
\value{
The output from \code{\link{print}}
}
\description{
My hello function
}
\examples{

hello("John")
\dontrun{
hello("Steve")
}
}
