% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Title}
\usage{
hello(x, multiply.by.vec = c(1:100), run.parallel = FALSE,
  run.parallel.cores = NULL)
}
\arguments{
\item{x}{Lorem Ipsum is simply dummy text of the printing and typesetting}

\item{multiply.by.vec}{ndustry. Lorem Ipsum has been the industry's standard dummy}

\item{run.parallel}{text ever since the 1500s, when an unknown printer took a gal}

\item{run.parallel.cores}{type and scrambled it to make a type specimen book. It}
}
\value{
ypesetting, remaining essentially unchanged. It was popularise
}
\description{
Title
}
