% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hola mundo}
\usage{
hello(x)
}
\arguments{
\item{x}{Persona a saludar}
}
\value{
La salida de \code{link{print}}
}
\description{
Función para saludar
}
\examples{
hello("Vivi")
\dontrun{
hello("Steve")
}


}
