% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of the person to say hi}
}
\value{
The output from \code{\link{print}}
}
\description{
prints hello world
}
\examples{
hello("mustu")
\dontrun{
hello("misu")
}
}
