% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My Hello World FUnction}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of the person to say hi to}
}
\value{
The output from \code{\link{print}}
}
\description{
My Hello World FUnction
}
\examples{
hello("john")
\dontrun{
hello("steve")
}
}
