% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World Function
:)}
\usage{
hello(x)
}
\arguments{
\item{x}{Name of person to say hi to}
}
\value{
(what is the value that is output form that funtion) Here, it is the output from \code{\link{print}}
}
\description{
Hello World Function
:)
}
\examples{
hello("Mary")
\dontrun{
hello("Steve")
}
}
