% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello function}
\usage{
hello(x)
}
\arguments{
\item{x}{the person to say hi to}
}
\value{
The output from \code{\link{print}}
}
\description{
A Function to say hello
}
\examples{
hello("John")
\dontrun{
hello("Steve")
}
}
