% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{Silly One}
\usage{
printer(r, x)
}
\arguments{
\item{r}{What you want in second column}

\item{x}{what you want in first column}
}
\value{
A tibble
}
\description{
Silly One
}
\examples{
printer(x=rnorm(5),r=rnorm(5))
}
