% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{My Lame printing function}
\usage{
printer(r, x)
}
\arguments{
\item{r}{What printed in the sexond column}

\item{x}{What printed in the first column}
}
\value{
A tibble
}
\description{
My Lame printing function
}
\examples{
printer(x = rnorm(5), r = rnorm(5))
}
