% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{Printer function}
\usage{
printer(r, x, y)
}
\arguments{
\item{r}{what you want in the second column}

\item{x}{what you want in the first column}
}
\value{
A tibble
}
\description{
Printer function
}
