% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{Silly Print Function}
\usage{
printer(r, x, y)
}
\arguments{
\item{r}{What you want in second column}

\item{x}{What you want in first column}

\item{y}{This is still another variable}
}
\value{
A tibble
}
\description{
Silly Print Function
}
\examples{
printer(x = rnorm(5), r = rnorm(5))
}
